pNFS/NFSv4: Try to return invalid layout in
Git-commit: 386b142945d3a8842c615421158f6d8565ef86e7
 pnfs_layout_process()
References: bsn#19
Patch-mainline: v5.10.16

[ Upstream commit 08bd8dbe88825760e953759d7ec212903a026c75 ]

If the server returns a new stateid that does not match the one in our
cache, then try to return the one we hold instead of just invalidating
it on the client side. This ensures that both client and server will
agree that the stateid is invalid.

Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.16-0026-pNFS-NFSv4-Try-to-return-invalid-layout-in-pnfs_layo.patch