scsi: hisi_sas: Remove deferred probe check in
From 4d287d8bae1f395b5e5d79bc9673dacab7975e36 Mon Sep 17 00:00:00 2001
 hisi_sas_v2_probe()
Git-commit: 4d287d8bae1f395b5e5d79bc9673dacab7975e36
Patch-mainline: v5.12-rc2
References: bsn#12

The platform_get_irq() check for -EPROBE_DEFER was to ensure that all the
steps to add the SCSI host are not done and then only to realise that the
probe needs to be deferred.

However, since there is now an earlier check for this in
hisi_sas_interrupt_preinit(), this check is superfluous and may be removed.

Link: https://lore.kernel.org/r/1611659068-131975-2-git-send-email-john.garry@huawei.com
Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.12-rc1-dontuse-scsi-hisi_sas-Remove-deferred-probe-check-in-hisi_sa.patch