xfs: refactor inode creation transaction/inode/quota
From f2f7b9ff62a28928f6fe2bd55cdb4d4b02ab7477 Mon Sep 17 00:00:00 2001
 allocation idiom
Git-commit: f2f7b9ff62a28928f6fe2bd55cdb4d4b02ab7477
Patch-mainline: v5.12-rc2
References: bsn#12
Modified-by-SEL: No

For file creation, create a new helper xfs_trans_alloc_icreate that
allocates a transaction and reserves the appropriate amount of quota
against that transction.  Replace all the open-coded idioms with a
single call to this helper so that we can contain the retry loops in the
next patchset.

This changes the locking behavior for non-tempfile creation slightly, in
that we now make the quota reservation without holding the directory
ILOCK.  While the dquots chosen for inode creation are based on the
directory state at a given point in time, the directory ILOCK was
released as soon as the dquot references are picked up.  Hence it was
never necessary to hold the directory ILOCK for the quota reservation.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.mainline/v5.12-rc2-xfs-refactor-inode-creation-transaction-inode-quota-.patch